Add a reftest.
Probably not the most reliable test to catch this (as it depends on
the font and such) but still fails here and I'd like WPT to have a test-case
for this. I'm looking into adding a WR test-case ATM, though fighting with
it way too much.

Differential Revision: https://phabricator.services.mozilla.com/D8438

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1494685
gecko-commit: d44ae1668a7020fe2f03a8d071876e25478d0b51
gecko-integration-branch: central
gecko-reviewers: kats